[NS] Fix issue with abstract lifecycled and the failed init handler
Execution of the failed init handler's destroy invocation potentially on
the same thread caused false positives in the tests here. The test was
completed successfully but if the lifecycle action happened to be on
another thread then the tests would fail.

Since we don't need the same thread behavior anymore it has been removed
and destroy will always happen after init completes as a separate action
on the lifecycle queue.

This exposed the bug where we were not resuming the lifecycle queue
after starting the destroy action, so we now resume the queue after
destroy is started.